backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "python-multipart>=0.0.9",
    "bcrypt>=4.0.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
